[OGUI-755] Investigate failure of ILG tests (#681)
* Splits test file into multiple ones to improve maintainability;
* Removes the retry option as it is creating unwanted scenarios in integration-tests
* Adds error logger for the dummy infologger server
* Update assert methods to recommended ones
